pyproject.toml
[project]
name = "job-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "beautifulsoup4>=4.14.2",
    "dotenv>=0.9.9",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "langchain>=1.0.3",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "langgraph>=1.0.2",
    "linkedin-scraper>=2.11.5",
    "openai>=2.7.1",
    "pydantic>=2.12.4",
    "sqlmodel>=0.0.27",
    "streamlit>=1.51.0",
    "streamlit-extras>=0.7.8",
    "uvicorn>=0.38.0",
    "watchdog>=6.0.0",
]
